// apps/web/src/features/poll/components/forms/poll-options-form/types.ts
export type DateOption = {
  type: "date";
  date: string;
};

export type TimeOption = {
  type: "timeSlot";
  start: string;
  end: string;
};

export type DateTimeOption = DateOption | TimeOption;

export interface DateTimePickerProps {
  title?: string;
  options: DateTimeOption[];
  date?: Date;
  onNavigate: (date: Date) => void;
  onChange: (options: DateTimeOption[]) => void;
  duration: number;
  onChangeDuration: (duration: number) => void;
  scrollToTime?: Date;
}